requirements.txt
streamlit>=1.42.0
Authlib>=1.3.2
requests>=2.31.0
pandas>=2.0.0
google-cloud-bigquery>=3.0.0
